+++ Old/run.py
@@ -3,8 +3,7 @@ import numpy as np
 import argparse
 from dataloader import load_extreme_data
 from trainOLD import train
-from network import Graph
-from communicators import CentralizedSGD, LSHCentralizedSGD
+from Other.communicators import CentralizedSGD, LSHCentralizedSGD
 from mlp import SparseNeuralNetwork
 from unpack import get_model_architecture, flatten_weights
 from mpi4py import MPI
