+++ Old/runOLD.py
@@ -3,7 +3,7 @@ import numpy as np
 import argparse
 from dataloader import load_extreme_data
 from Old.train import train
-from communicators import CentralizedSGD, LSHCentralizedSGD
+from Other.communicators import CentralizedSGD, LSHCentralizedSGD
 from mlp import SparseNeuralNetwork
 from Old.unpack import get_model_architecture, flatten_weights
 from mpi4py import MPI
